+++ plugins/ucc_chat/lib/ucc_chat_web/flex_bar/tab/members_list.ex
@@ -189,7 +189,7 @@ defmodule UccChatWeb.FlexBar.Tab.MembersList do
 
   def set_mute(socket, sender) do
     username = select socket, data: "username", from: ".user-view[data-username]"
-    Logger.warn "username: #{inspect username}, sender: #{inspect sender}"
+    # Logger.warn "username: #{inspect username}, sender: #{inspect sender}"
     user = UcxUcc.Accounts.get_by_user username: username
     channel_id = socket.assigns.channel_id
     current_user = UcxUcc.Accounts.get_user socket.assigns.user_id, preload: [:roles, user_roles: :role]
@@ -206,7 +206,7 @@ defmodule UccChatWeb.FlexBar.Tab.MembersList do
 
   def unset_mute(socket, sender) do
     username = select socket, data: "username", from: ".user-view[data-username]"
-    Logger.warn "username: #{inspect username}, sender: #{inspect sender}"
+    # Logger.warn "username: #{inspect username}, sender: #{inspect sender}"
     user = UcxUcc.Accounts.get_by_user username: username
     channel_id = socket.assigns.channel_id
     current_user = UcxUcc.Accounts.get_user socket.assigns.user_id, preload: [:roles, user_roles: :role]
@@ -223,7 +223,7 @@ defmodule UccChatWeb.FlexBar.Tab.MembersList do
 
   def set_owner(socket, sender) do
     username = select socket, data: "username", from: ".user-view[data-username]"
-    Logger.warn "username: #{inspect username}, sender: #{inspect sender}"
+    # Logger.warn "username: #{inspect username}, sender: #{inspect sender}"
     user = UcxUcc.Accounts.get_by_user username: username
     channel_id = socket.assigns.channel_id
     current_user = UcxUcc.Accounts.get_user socket.assigns.user_id, preload: [:roles, user_roles: :role]
@@ -240,7 +240,7 @@ defmodule UccChatWeb.FlexBar.Tab.MembersList do
 
   def unset_owner(socket, sender) do
     username = select socket, data: "username", from: ".user-view[data-username]"
-    Logger.warn "username: #{inspect username}, sender: #{inspect sender}"
+    # Logger.warn "username: #{inspect username}, sender: #{inspect sender}"
     user = UcxUcc.Accounts.get_by_user username: username
     channel_id = socket.assigns.channel_id
     current_user = UcxUcc.Accounts.get_user socket.assigns.user_id, preload: [:roles, user_roles: :role]
@@ -257,7 +257,7 @@ defmodule UccChatWeb.FlexBar.Tab.MembersList do
 
   def set_moderator(socket, sender) do
     username = select socket, data: "username", from: ".user-view[data-username]"
-    Logger.warn "username: #{inspect username}, sender: #{inspect sender}"
+    # Logger.warn "username: #{inspect username}, sender: #{inspect sender}"
     user = UcxUcc.Accounts.get_by_user username: username
     channel_id = socket.assigns.channel_id
     current_user = UcxUcc.Accounts.get_user socket.assigns.user_id, preload: [:roles, user_roles: :role]
@@ -274,7 +274,7 @@ defmodule UccChatWeb.FlexBar.Tab.MembersList do
 
   def unset_moderator(socket, sender) do
     username = select socket, data: "username", from: ".user-view[data-username]"
-    Logger.warn "username: #{inspect username}, sender: #{inspect sender}"
+    # Logger.warn "username: #{inspect username}, sender: #{inspect sender}"
     user = UcxUcc.Accounts.get_by_user username: username
     channel_id = socket.assigns.channel_id
     current_user = UcxUcc.Accounts.get_user socket.assigns.user_id, preload: [:roles, user_roles: :role]
@@ -328,7 +328,6 @@ defmodule UccChatWeb.FlexBar.Tab.MembersList do
         ]
     socket.endpoint.broadcast! CC.chan_room <> socket.assigns.room,
       "update:flex-button", %{username: user.username, html: html, button: "mute-unmute"}
-    socket
   end
 
 end